+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The types of changes are:
 * Config Wizard: Added a column selector to the scan results page of the config wizard [#1590](https://github.com/ethyca/fides/pull/1590)
 * Config Wizard: Flow for runtime scanner option [#1640](https://github.com/ethyca/fides/pull/1640)
 * Access support for Twilio Conversations API [#1520](https://github.com/ethyca/fides/pull/1520)
+* Message Config: Adds Twilio Email/SMS support [#1519](https://github.com/ethyca/fides/pull/1519)
 
 ### Changed